// frontend/src/app/core/services/api/model/scoringRuleDto.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BoatClass } from './boatClass';


/**
 * Data Transfer Object for scoring rule
 */
export interface ScoringRuleDto { 
    uuid?: string;
    scoringSchemeId?: string;
    placement?: number;
    basePoints?: number;
    boatClass?: BoatClass;
    coefficient?: number;
    createdAt?: string;
    modifiedAt?: string;
}
export namespace ScoringRuleDto {
}