// frontend/src/app/core/services/api/model/scoringRuleRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request payload for creating or updating a scoring rule
 */
export interface ScoringRuleRequest { 
    scoringSchemeId: string;
    placement: number;
    basePoints: number;
}