// frontend/src/app/core/services/api/model/scoringSchemeBoatCoefficientDto.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BoatClass } from './boatClass';


/**
 * Data Transfer Object for scoring scheme boat coefficient
 */
export interface ScoringSchemeBoatCoefficientDto { 
    uuid?: string;
    scoringSchemeId?: string;
    boatClass?: BoatClass;
    coefficient?: number;
    createdAt?: string;
    modifiedAt?: string;
}
export namespace ScoringSchemeBoatCoefficientDto {
}